pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "runningman"
description = "Runtime Manager, i.e. a running-man(ager), for setting up a set of service like processes that are asynchronously managed by triggers and fed data trough queues."
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    {name = "Daniel Kastinen", email = "daniel.kastinen@irf.se"},
]
maintainers = [
    {name = "Daniel Kastinen", email = "daniel.kastinen@irf.se"},
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.7"
dependencies = [
    "croniter",
    "pyzmq",
]
dynamic = ["version"]

[project.scripts]
runningman = "runningman.cli:main"
rain-register = "rain.cli:register_cli"
rain-server = "rain.cli:server_cli"
rain-trigger = "rain.cli:trigger_cli"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "runningman.version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
develop = [
    "pytest",
    "flake8",
    "wheel",
    "build",
    "twine",
    "numpydoc",
    "black",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
]

[project.urls]
Documentation = "https://danielk.developer.irf.se/runningman/"
Repository = "https://github.com/danielk333/runningman"

[tool.black]
line-length = 100
target_version = ['py35', 'py36', 'py37', 'py38', 'py39', 'py310']
